Major Changes & Additions
Inspired by the Christina Hunger's work to use an AAC button board to teach her dog to talk. (https://hungerforwords.com) This software plays audio clips based on buttons pressed. It runs as a simple service in the background on a Raspberry Pi Desktop. The Raspberry Pi interfaces with the buttons using two ( or more) USB "Zero-Delay" arcade button controllers for a total of 24+ buttons on one sound board.